+++ packages/client/app/(withLayout)/question/[id]/page.tsx
@@ -1,5 +1,5 @@
-import { QuestionDetailInformation } from '@widgets/QuestionDetailInformation/'
+import { QuestionInformation } from '@widgets/QuestionDetailInformation/'
 
 export default function QuestionDetailPage() {
-   return <QuestionDetailInformation />
+   return <QuestionInformation />
 }
