+++ packages/client/src/widgets/QuestionDetailInformation/ui/questionInformation.css.ts
